+++ lib/date_time_basics.dart
@@ -54,4 +54,65 @@ extension DateTimeBasics on DateTime {
   /// Delegates to [DateTime]'s built-in comparison methods and therefore obeys
   /// the same contract.
   bool isAtOrAfter(DateTime other) => isAtSameMomentAs(other) || isAfter(other);
+
+  /// Copies a [DateTime], overriding specified values.
+  ///
+  /// A UTC [DateTime] will remain in UTC; a local [DateTime] will remain local.
+  DateTime copyWith({
+    int? year,
+    int? month,
+    int? day,
+    int? hour,
+    int? minute,
+    int? second,
+    int? millisecond,
+    int? microsecond,
+  }) {
+    return (isUtc ? DateTime.utc : DateTime.new)(
+      year ?? this.year,
+      month ?? this.month,
+      day ?? this.day,
+      hour ?? this.hour,
+      minute ?? this.minute,
+      second ?? this.second,
+      millisecond ?? this.millisecond,
+      microsecond ?? this.microsecond,
+    );
+  }
+
+  /// Adds a specified number of days to this [DateTime].
+  ///
+  /// Unlike `DateTime.add(Duration(days: numberOfDays))`, this adds calendar
+  /// days and not 24-hour increments.  When possible, it therefore leaves the
+  /// time of day unchanged if a DST change would occur during the time
+  /// interval. (The returned time can still be different from the original if
+  /// it would be invalid for the returned date.)
+  DateTime addCalendarDays(int numberOfDays) =>
+      copyWith(day: day + numberOfDays);
+
+  /// Returns the number of calendar days till the specified date.
+  ///
+  /// Returns a negative value if the specified date is in the past.  Ignores
+  /// the time of day.
+  ///
+  /// Example:
+  /// ```
+  /// DateTime(2020, 12, 31).calendarDaysTill(2021, 1, 1); // 1
+  /// DateTime(2020, 12, 31, 23, 59).calendarDaysTill(2021, 1, 1); // 1
+  /// ```
+  ///
+  /// This function intentionally does not take a [DateTime] argument to:
+  /// * More clearly indicate that it does not take time of day into account.
+  /// * Avoid potential problems if one [DateTime] is in UTC and the other is
+  ///   not.
+  int calendarDaysTill(int year, int month, int day) {
+    // Discard the time of day, and perform all calculations in UTC so that
+    // Daylight Saving Time adjustments are not a factor.
+    //
+    // Note that this intentionally isn't the same as `toUtc()`; we instead
+    // want to treat this `DateTime` object *as* a UTC `DateTime`.
+    final startDay = DateTime.utc(this.year, this.month, this.day);
+    final endDay = DateTime.utc(year, month, day);
+    return (endDay - startDay).inDays;
+  }
 }